backend/ai_service/requirements.txt
flask
transformers
Pillow
torch
